airflow_data_pipeline/requirements.txt
passlib
google-cloud-storage
python-decouple
apache-airflow[gcp]==2.4.2

